refactor: update client components, improve agent metrics, and add compliance utilities
