+++ wazuh-agent-status-rust-server/src/status_provider/linux.rs
@@ -18,7 +18,6 @@ pub struct LinuxStatusProvider {
 impl LinuxStatusProvider {
     pub fn new(paths: AgentPaths) -> Self {
         let mut sys = System::new();
-        // Initial refresh so we have something for first poll
         sys.refresh_all();
         Self {
             paths,
@@ -26,9 +25,6 @@ impl LinuxStatusProvider {
         }
     }
 
-    /// Determine whether the `wazuh-agentd` process is alive by checking the
-    /// process list via `sysinfo`. This avoids lock file race conditions
-    /// inherent in calling `wazuh-control status`.
     fn is_agent_running(&self) -> bool {
         if let Ok(mut sys) = self.sys.lock() {
             sys.refresh_processes(sysinfo::ProcessesToUpdate::All, true);
@@ -129,44 +125,12 @@ impl StatusProvider for LinuxStatusProvider {
             .lock()
             .map_err(|_| ServerError::PlatformError("Failed to lock system metrics".to_string()))?;
 
-        // Refresh all processes to find the Wazuh ones
         sys.refresh_processes(sysinfo::ProcessesToUpdate::All, true);
         sys.refresh_memory();
-        sys.refresh_cpu_all(); // Ensure core information is fresh for delta calculation
+        sys.refresh_cpu_all();
 
-        let mut total_cpu: f32 = 0.0;
-        let mut total_rss: u64 = 0;
-        let mut found_names = Vec::new();
-
-        for process in sys.processes().values() {
-            let name = process.name().to_string_lossy();
-            if crate::status_provider::UNIX_AGENT_PROCESSES.contains(&name.as_ref()) {
-                let p_cpu = process.cpu_usage();
-                total_cpu += p_cpu;
-                total_rss += process.memory();
-                found_names.push(format!("{} ({:.1}%)", name, p_cpu));
-            }
-        }
-
-        let cpu_count = sys.cpus().len() as f32;
-        let cpu_usage = if !found_names.is_empty() && cpu_count > 0.0 {
-            total_cpu / cpu_count
-        } else {
-            0.0
-        };
-
-        let total_memory = sys.total_memory();
-        let memory_usage = if total_memory > 0 {
-            total_rss as f32 / total_memory as f32
-        } else {
-            0.0
-        };
-
-        Ok(crate::models::SystemMetrics {
-            cpu_usage,
-            memory_usage,
-            total_memory,
-            used_memory: total_rss,
-        })
+        Ok(crate::status_provider::unix::collect_unix_system_metrics(
+            &sys,
+        ))
     }
 }
